asdasd

+++ database/migrations/2020_03_26_061525_create_units_table.php
@@ -16,7 +16,7 @@ class CreateUnitsTable extends Migration
         Schema::create('units', function (Blueprint $table) {
             $table->bigIncrements('unit_id');
             $table->string('unit_no',10);
-            $table->unsignedInteger('unit_unit_owner_id');
+            $table->bigIncrements('unit_unit_owner_id');
             $table->integer('floor_no');
             $table->integer('beds');
             $table->float('monthly_rent', 8,2);
