+++ database/migrations/2020_04_27_093931_create_billings_table.php
@@ -15,7 +15,7 @@ class CreateBillingsTable extends Migration
     {
         Schema::create('billings', function (Blueprint $table) {
             $table->bigIncrements('billing_id');
-            $table->unsignedInteger('billing_tenant_id');   
+            $table->bigIncrements('billing_tenant_id');   
             
             $table->date('billing_date');
             $table->string('billing_desc');
